Deal with STAC items that have only partial proj data (#19)
Before this change we assumed the following

- If an Item declares proj extension, then it includes shape/transform data
- And hence only bands with proj metadata set should be considered as valid

However it is possible that only CRS information is supplied on an Item, in
this case we would find no data bands at all and then fail in unexpected way.

The change here is to retain the same rule, but with a minor twist

- If there are any bands with proj data, take only those
- Else it is as if there were no proj extension data at all

This is a little bit sub-optimal, as it would be nice to propagate CRS
information from STAC to Dataset even when shape/transform information is missing.